refactor(ui): decompose AI config God Component into sub-components
Breaks the 1213-line page.tsx into focused sub-components:

- page.tsx (186 lines): thin orchestrator with state + tabs
- components/tab-geral.tsx: provider, API key, model, temperature, limits
- components/tab-whatsapp.tsx: WhatsApp toggle, persona, keywords, iterations
- components/tab-email.tsx: email toggle, persona, signature
- components/tab-consumo.tsx: usage dashboard with channel/model breakdowns
- components/tab-simulador.tsx: AI response simulation sandbox
- components/types.ts: shared constants and type re-exports
- components/index.ts: barrel export

Zero behavior changes — same UI, same state flow, same server actions.

Closes #149
