+++ erp/src/app/(app)/configuracoes/ai/components/index.ts
@@ -1,0 +1,7 @@
+export { TabGeral } from "./tab-geral";
+export { TabWhatsApp } from "./tab-whatsapp";
+export { TabEmail } from "./tab-email";
+export { TabConsumo } from "./tab-consumo";
+export { TabSimulador } from "./tab-simulador";
+export { PROVIDERS, DEFAULT_CONFIG } from "./types";
+export type { AiConfigData } from "./types";
